Minor bug fixes in data packages + improved info schema demo to connect to mysql and postgres as well as sql server..

+++ Extenso.Data/SqlClient/SqlConnectionExtensions.cs
@@ -110,7 +110,7 @@ AND T.CONSTRAINT_SCHEMA = @SchemaName";
                             //}
 
                             if (!reader.IsDBNull(3))
-                            { columnInfo.MaximumLength = reader.GetInt32(3); }
+                            { columnInfo.MaximumLength = reader.GetInt64(3); }
 
                             if (!reader.IsDBNull(4))
                             {
